requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pandas==2.2.3
pytest==8.3.3
httpx==0.27.2
scikit-learn==1.5.2
xgboost==2.1.1
sentence-transformers==3.0.1
numpy
PyYAML==6.0.2